+++ demos/manual/src/main/java/is/codion/manual/swing/common/ui/component/table/FilterTableDemo.java
@@ -45,7 +45,8 @@ final class FilterTableDemo {
 						FilterTableColumn.builder(Person.AGE, 1).build());
 
 		FilterTable<Person, String> table =
-						FilterTable.builder(tableModel)
+						FilterTable.builder()
+										.model(tableModel)
 										.columns(columns)
 										.doubleClick(Control.command(() ->
 														tableModel.selection().item().optional()