+++ tools/monitor/ui/src/main/java/is/codion/tools/monitor/ui/ServerMonitorPanel.java
@@ -239,7 +239,8 @@ public final class ServerMonitorPanel extends JPanel {
 
 	private JPanel createOperationPanel() {
 		FilterTable<DomainOperation, OperationColumns.Id> table =
-						FilterTable.builder(model.operationTableModel())
+						FilterTable.builder()
+										.model(model.operationTableModel())
 										.columns(createOperationColumns())
 										.autoResizeMode(JTable.AUTO_RESIZE_ALL_COLUMNS)
 										.build();
@@ -258,7 +259,8 @@ public final class ServerMonitorPanel extends JPanel {
 
 	private JPanel createReportPanel() {
 		FilterTable<DomainReport, ReportColumns.Id> table =
-						FilterTable.builder(model.reportTableModel())
+						FilterTable.builder()
+										.model(model.reportTableModel())
 										.columns(createReportColumns())
 										.autoResizeMode(JTable.AUTO_RESIZE_ALL_COLUMNS)
 										.build();
@@ -280,7 +282,8 @@ public final class ServerMonitorPanel extends JPanel {
 
 	private JPanel createEntityPanel() {
 		FilterTable<DomainEntityDefinition, DomainColumns.Id> table =
-						FilterTable.builder(model.domainTableModel())
+						FilterTable.builder()
+										.model(model.domainTableModel())
 										.columns(createDomainColumns())
 										.autoResizeMode(JTable.AUTO_RESIZE_ALL_COLUMNS)
 										.build();